+++ packages/ui/src/components/Form/NumberFormItem/index.tsx
@@ -2,14 +2,16 @@ import React from 'react';
 import { Control, FieldValues, UseFormReturn } from 'react-hook-form';
 import { FormControl, FormField, FormItem, FormLabel, FormMessage } from '@ui/src/Form';
 import { Input } from '@ui/src/Input';
-import { TextBody1 } from '../Text';
+import { TextBody1 } from '../../Text';
 
-type InputFormItemProps = {
+type NumberInputFormItemProps = {
   control: Control<any, any>;
   name: string;
   label?: string | React.ReactNode;
   placeholder: string;
   required?: boolean;
+  unit?: string;
+  range?: [number, number];
 };
 /**
  * Renders a form item for an input field with optional required label.
@@ -20,15 +22,29 @@ type InputFormItemProps = {
  * @param {string} props.placeholder - The placeholder for the input field.
  * @return {JSX.Element} The rendered input form item.
  */
-const InputFormItem = ({ required = false, ...props }: InputFormItemProps) => {
+const NumberInputFormItem = ({ range, unit, required = false, ...props }: NumberInputFormItemProps) => {
   return (
     <FormField
       control={props.control}
       name={props.name}
-      rules={{ required: required }}
+      rules={{ required: required, min: 0, max: 60 }}
       render={({ field }) => {
+        const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+          let value = e.target.value;
+          if (range) {
+            if (parseInt(e.target.value) >= range[0] && parseInt(e.target.value) <= range[1]) {
+              value = e.target.value;
+            } else if (parseInt(e.target.value) > range[1]) {
+              value = String(range[1]);
+            } else {
+              value = '';
+            }
+          }
+          field.onChange(value);
+        };
+
         return (
-          <FormItem>
+          <FormItem className="relative">
             <FormLabel>
               <TextBody1>
                 {props.label}
@@ -36,8 +52,16 @@ const InputFormItem = ({ required = false, ...props }: InputFormItemProps) => {
               </TextBody1>
             </FormLabel>
             <FormControl>
-              <Input className="bg-gray-50 pl-[14px] h-[50px]" placeholder={props.placeholder} {...field} />
+              <Input
+                className="bg-gray-50 h-[56px] text-2xl pl-[14px]"
+                placeholder={props.placeholder}
+                {...field}
+                onChange={handleChange}
+                type="number"
+              />
             </FormControl>
+            {unit && field.value && <span className="absolute bottom-[16px] left-[35px]">{unit}</span>}
+
             <FormMessage />
           </FormItem>
         );
@@ -46,4 +70,4 @@ const InputFormItem = ({ required = false, ...props }: InputFormItemProps) => {
   );
 };
 
-export default InputFormItem;
+export default NumberInputFormItem;
